Added new interaction elements (buy, fav)
- Added new endpoint for multiple actions
    - POST /v1/{app_id}/index/{index_id}/interaction/{interaction_type}

- Refactored the tests to be MUCH faster (http instead of curl in most
cases)
- When the user nor the IP is set on query, the search is not registered

+++ Plugin/DBAL/DependencyInjection/DBALPluginExtension.php
@@ -193,7 +193,7 @@ class DBALPluginExtension extends BaseExtension
                 $encryptIsEnabled
             ),
 
-            'apisearch_plugin.dbal.loop_push_interval' => \intval(Env::get(
+            'apisearch_plugin.dbal.loop_push_interval' => \floatval(Env::get(
                 'DBAL_LOOP_PUSH_INTERVAL',
                 $config['loop_push_interval']
             )),
